Bugfix: when extracting the allocator held in the `Box<T,A>`, must replace it with the dropped flag marker (at least for now; stack local drop flags should fix this).
Added an explicit drop of the extracted allocator to remind the reader
that it happens. :)

Removed (unnecessary) attempt to mark the pointer to the value as
dropped.

Added some instrumentation to `<Box as Drop>::drop`, though TBH I do
not think this actually helped me track down the core issue in this
particular case. Anyway I figure I'll remove all that instrumentation
later.